Added check for terraform destroy to bucket deletion (#60)
* Added check for terraform destroy to bucket deletion
* Added missing readme and actions descritpion
* README typo

+++ operations/deployment/terraform/bitops.after-deploy.d/delete-tf-state-bucket.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-if [[ $TF_STATE_BUCKET_DESTROY = true ]] ; then
+if [[ $TERRAFORM_DESTROY = true ]] && [[ $TF_STATE_BUCKET_DESTROY = true ]]; then 
   echo "Destroying TF State S3 bucket --> $TF_STATE_BUCKET"
   aws s3 rb s3://$TF_STATE_BUCKET --force
 else
